split hassio.ts
