+++ hassio/src/ingress-view/hassio-ingress-view.ts
@@ -9,11 +9,11 @@ import {
   css,
 } from "lit-element";
 import { HomeAssistant, Route } from "../../../src/types";
+import { createHassioSession } from "../../../src/data/hassio/supervisor";
 import {
-  createHassioSession,
   HassioAddonDetails,
   fetchHassioAddonInfo,
-} from "../../../src/data/hassio";
+} from "../../../src/data/hassio/addon";
 import "../../../src/layouts/hass-loading-screen";
 import "../../../src/layouts/hass-subpage";
