+++ src/data/hassio/common.ts
@@ -1,0 +1,7 @@
+export interface HassioResponse<T> {
+  data: T;
+  result: "ok";
+}
+
+export const hassioApiResultExtractor = <T>(response: HassioResponse<T>) =>
+  response.data;
